move cli code to cli subdirectory
